requirements.txt
cachetools==3.1.1
certifi==2021.5.30
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Flask==2.0.1
gast==0.2.2
grpcio==1.27.2
gunicorn==20.1.0
h5py==2.10.0
idna==2.9
imageio==2.9.0
importlib-metadata==1.5.0
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.17.0
Jinja2==3.0.1
joblib==0.15.1
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==1.2.6
jupyterlab-server==1.1.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.1.1
MarkupSafe==2.0.1
matplotlib==3.2.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
notebook==6.0.3
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
pandas==1.0.4
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.7.0
pickleshare==0.7.5
Pillow==7.1.2
pycparser==2.20
Pygments==2.6.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyreadline==2.1
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
rsa==4.0
scikit-image==0.17.2
six==1.14.0
tensorboard==2.0.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow-cpu==2.3.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.0.0
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==2.0.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
zipp==3.1.0